Fix test setup to enable running the tests against the main system.
This means declaring an explicit dependency on the RT system (it's in
quicklisp) from the test system, moving the test system .asd to the
top-level where asdf will find it, having TEST-OP on the test system call
RT:DO-TESTS, and finally having the main system chain TEST-OP to the test
system.

The end result is that tests for py-configparser can be run by using e.g.
(ASDF:TEST-SYSTEM '#:py-configparser).